% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{what_time}
\alias{what_time}
\title{Indicate what time is it}
\usage{
what_time(language = "es")
}
\arguments{
\item{language}{string indicating the language}
}
\value{
a string with a sentence
}
\description{
details what time is it
}
\examples{
what_time()
what_time(language = "es")



}
